+++ app/core/models/order_items.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import TYPE_CHECKING
 
 from sqlalchemy import ForeignKey, Integer, Numeric
@@ -13,7 +14,7 @@ class OrderItem(Base):
     order_id: Mapped[int] = mapped_column(ForeignKey("orders.id"), nullable=False)
     nomenclature_id: Mapped[int] = mapped_column(ForeignKey("nomenclatures.id"), nullable=False)
     quantity: Mapped[int] = mapped_column(Integer, nullable=False)
-    price_at_purchase: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False)
+    price_at_purchase: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False)
 
     order: Mapped["Order"] = relationship(back_populates="items")
     nomenclature: Mapped["Nomenclature"] = relationship()
